order_bot/requirements.txt
# Discord Bot
discord.py>=2.3.2

# Web Framework untuk Webhook
Flask>=3.0.0

# PostgreSQL Adapter (for shared database)
psycopg2-binary>=2.9.9

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Encryption (for stock codes)
cryptography>=41.0.0

# Utilities
asyncio

# Optional: For SQLite mode
# No additional packages needed
